fixed all routes, deleted checkOwner script

+++ src/routes/Users.ts
@@ -3,13 +3,12 @@ import passport from 'passport'
 import express from 'express' 
 import { User } from '../drizzle/schema/user'
 import { createUser, deleteUser, getUserByUsername } from '../drizzle/actions/user'
-import '../Authentification/pasport_jwt'
 
 
 const router = express.Router()
-// 
 
 router.get('/getUserFromJwt', passport.authenticate('jwt', {session: false}), async (req, res) => {
+    
     const user = {id: req.user.id, username: req.user.username, email: req.user.email}
     res.status(200).json({success: true, user: user, msg: 'passport authentification went well, nicuuuuuuuuuuuu'})
 })
@@ -24,24 +23,25 @@ router.post('/login', async(req, res) => {
         
         const isValid = verifyPassword(password, user.password, user.salt)
         if (isValid){
-            const jwt = issueJWT(user)
+            const jwt = issueJWT(user.id)
             res.cookie('jwt', jwt, {maxAge:86400000, httpOnly: true})
             res.cookie('jwtExist', true, {maxAge:86400000})
             res.status(200).json({success: true, msg:'User successfully logged in', user: user})
         } else {
             res.status(401).json({success: false, msg:'password is incorrect'})
         }
-    }).catch((err) => res.json({msg:err}))
+    }).catch((err) => {
+        res.json({msg: err.message})
+    })
 })
 
 router.post('/createUser', async(req, res) => {
     const user = req.body
-    console.info('user=======>',user);
+    console.info('user from request=======>',user);
     const {salt, hash} = genPassword(user.password)
     user.password = hash
     user.salt = salt
 
-    console.info('user=======>',user);
     await createUser(user)
         .then((userId: bigint) => {
             const jwt = issueJWT(userId)
@@ -52,13 +52,18 @@ router.post('/createUser', async(req, res) => {
     // res.send('The error ocurred during creating(posting) user entity into the table')
 });
 
-router.post('/deleteUser', passport.authenticate('jwt', {session: false}), async (req, res) => {
-    const userId = req.user?.id
-    await deleteUser(userId)
-    res.status(200).json({success: true, msg: 'User successfully deleted'})
+router.delete('/deleteUser', passport.authenticate('jwt', {session: false}), async (req, res) => {
+    const userId = (req.user as User).id
+    if(userId){
+        await deleteUser(userId)
+        res.status(200).json({success: true, msg: 'User successfully deleted'})
+    }else{
+        res.status(401).json({success: false, msg: 'Error: user not found or anauthorized'})
+    }
+    
 })
 
-router.get('/logout', async(req, res) => {
+router.get('/logout', passport.authenticate('jwt', {session: false}), async(req, res) => {
     res.cookie('jwt', false, {maxAge:0, httpOnly: true})
     res.cookie('jwtExist', false, {maxAge:0})
     res.status(200).send('User successfuly logged out')
